backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
pydantic==2.9.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
PyGithub==2.4.0
beautifulsoup4==4.12.3
trafilatura==1.12.1
ollama>=0.4.0
scikit-learn==1.5.2
pandas==2.2.2
pymupdf>=1.24.0
python-multipart>=0.0.9
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
